requirements.txt
dearpygui==1.9.0
Pillow==9.5.0
